[~] refactor code, delete swagger logic from routes.
- fix some docs issues
