+++ effectiveOfficeBackend/src/main/kotlin/office/effective/common/swagger/SwaggerDocument.kt
@@ -1,0 +1,4 @@
+package office.effective.common.swagger
+
+public object SwaggerDocument {
+}
